modules/location_service/location_service_listen/requirements.txt
grpcio 
grpcio-tools 
kafka-python 

